% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_cor_tab.R
\name{apa_cor_table}
\alias{apa_cor_table}
\title{APA Correlation Table}
\usage{
apa_cor_table(dat, dig = 2, stars = TRUE, msd = TRUE, minmax = FALSE)
}
\arguments{
\item{dat}{The data.frame of continuous variables you'd like to make a table for}

\item{dig}{How many digits should values round to? Defaults to 2}

\item{stars}{Logical. Should significance stars be included? Defaults to TRUE}

\item{msd}{Logical. Should the mean and standard deviation be included? Defaults to TRUE.}

\item{minmax}{Logical. Should the minimum and maximum be included? Defaults to FALSE.}
}
\description{
This is a general function for building typical tables that I see in reports. Similar to \code{apa.cor.table} from the \code{apaTables} package.
}
\details{
Calculates a base correlation matrix, then adds leading columns as requested. Rounding is conducted via \code{myStuff::fr} and stars are calculated via \code{demfunks::sig_stars}.
}
\examples{
apa_cor_table(iris[,-5])

}
\seealso{
\code{\link[apaTables:apa.cor.table]{apaTables::apa.cor.table()}}
}
\author{
Michael Floren
}
