% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_row.R
\name{cat_row}
\alias{cat_row}
\title{Catgorical Descriptives as Row}
\usage{
cat_row(x, header_name = NA, ...)
}
\arguments{
\item{x}{The vector with categorical data}

\item{header_name}{Character. The name to include at the top of the listing.}

\item{...}{Other arguments to pass to \code{catf}.}
}
\description{
This function is a wrapper for \code{catf} that presents the categorical entries as rows (with a potential header). This is useful when creating categorical tables.
}
\details{
This function was created to ease the inclusion of \code{catf} results into report table formats.
}
\examples{
library(flextable)
library(dplyr)

temp <- rbind(cat_row(ChickWeight$Diet, header_name = "Diets"),
              cat_row(ChickWeight$Time, header_name = "Times"))
              
cat_tab <- data.frame(Characteristics=rownames(temp), "Values"=temp)

rows_to_indent <- which(!is.na(temp))

flextable(cat_tab) \%>\%
  set_header_labels(Values="Values") \%>\%
  align(j=1, align="left", part="all") \%>\%
  align(j=2, align="center", part="all") \%>\% 
  italic(i=1, part="head") \%>\%
  padding(i=rows_to_indent, j=1, padding.left = 25) \%>\%
  font(fontname="Times New Roman", part="all") \%>\%
  fontsize(size=12, part="all") \%>\%
  autofit()

}
\seealso{
Other categorical data functions: 
\code{\link{catf}()},
\code{\link{ind_catf}()},
\code{\link{sub_catf}()}
}
\author{
Michael Floren
}
\concept{categorical data functions}
