% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clpse_sata.R
\name{clpse_sata}
\alias{clpse_sata}
\title{Collapse Select All That Apply}
\usage{
clpse_sata(
  x,
  multiple_text = "Multiple",
  mark_scheme = "anyvsNA",
  none_text = NA
)
}
\arguments{
\item{x}{A data.frame including the select-all-that-apply group to be condensed}

\item{multiple_text}{What text should be used when multiple categories have been selected? Default is "Multiple".}

\item{none_text}{What text should be used when none are selected? Defaults to \code{NA}.}

\item{mark_sceme}{Currently not used. What marking scheme should be used to assess which entries are "selected" or "marked"?}
}
\description{
This function is used to collapse select all that apply questions into a single column (with options for how to address when multiple options are selected). Useful when some demographic information (e.g., race) is listed as a select-all-that-apply, but is analyzed as a single categorical variable.
}
\details{
For each entry, assessments are made on each row if multiple selections have been made, or if no selections have been made. Current implementation checks vs NA to see (i.e., assumes any text is a "selection", and all unselected entries are marked "NA" (\emph{not} 0, or something else)). For rows with multiple, use the multiple text. For rows with none, marks as \code{NA} (default) or use \code{none_text}. For rows with a single entry, use the column name of the row that has that entry (i.e., column names should be pretty).
}
\examples{
set.seed(1)
df <- data.frame(white = sample(c(NA,1), 100, replace=TRUE),
                 black = sample(c(NA,1), 100, replace=TRUE))
clpse_sata(df)

}
\author{
Michael Floren
}
