% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contf.R
\name{contf}
\alias{contf}
\title{Continuous Descriptives Base Function}
\usage{
contf(
  x,
  na.rm = TRUE,
  dig = 2,
  morm = "mean",
  supp = "match",
  alpha = 0.05,
  n = TRUE,
  quart_vals = c(0.25, 0.75),
  quart_type = 2
)
}
\arguments{
\item{x}{The numeric vector to be used in calculating the descriptives.}

\item{na.rm}{Should missing values be removed? Defaults to TRUE}

\item{dig}{How many digits should values round to? Defaults to 2}

\item{morm}{Which measure of central tendency would you like? Options are "mean" or "median"}

\item{supp}{Which supplementary statistic(s) would yu like to be provided? Options are "none", "match", "sd", "ci", "range", and "quartiles". "match" is default, and will provide "sd" for mean and "quartiles" for median.}

\item{alpha}{What alpha level would you like to set for the CI? Defaults to 0.05 (only used for confidence intervals)}

\item{n}{Should sample size be provided? Defaults to TRUE.}

\item{quart_vals}{Numeric vector of length 2, passed to \code{quantile} when quartiles are requested. What proportions should be used when calculating quartiles? Defaults to c(0.25, 0.75).}

\item{quart_type}{What type of quantiles should be used in the quartile calculation (passed to \code{quantile}).}
}
\description{
This function generates a short summary of descriptive statistics for a single numeric vector.Additional arguments are allowed to specify desired measure of central tendancy, desired additional information (e.g., SD, quartiles), and whether or not sample size is reported.
}
\details{
Calculations use R standard functions. Quantiles is used for the calculation of quartiles.
}
\examples{
contf(ChickWeight$weight)
contf(ChickWeight$weight, dig=3, supp="ci", alpha=0.10)
contf(ChickWeight$weight, dig=0, morm="median", quart_vals = c(0.05, 0.95), quart_type=5, n=FALSE)

}
\seealso{
Base calculation functions of \code{\link[=mean]{mean()}}, \code{\link[=median]{median()}}, \code{\link[=sd]{sd()}}, and \code{\link[=quantile]{quantile()}}. Formatting and rounding is done using \code{\link[=fr]{fr()}}.
}
\author{
Michael Floren
}
