% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_catf.R
\name{ind_catf}
\alias{ind_catf}
\title{Categorical Descriptives Individual Base Function}
\usage{
ind_catf(
  x,
  dig = 1,
  na_in_denom = FALSE,
  which_in_paren = "p",
  include_perc_sign = TRUE,
  ...
)
}
\arguments{
\item{x}{The binary vector (0's and 1's) with data on the categorical variable to be used in calculating the descriptives.}

\item{dig}{How many digits should values round to? Defaults to 1 (e.g., 99.1\%).}

\item{na_in_denom}{Should NA's be included in the denominator when calculating the percentage? Note that NA's will not be included in the sum. Default is FALSE.}

\item{which_in_paren}{Which should be in parentheses? Arguments are "n" and "p". Defaults to "p"}

\item{include_perc_sign}{Should the percent sign be included? Defaults to FALSE.}

\item{...}{Additional arguments for \code{table} (e.g., \code{exclude=NULL} to include NA's)}
}
\description{
This function generates a short summary of descriptive statistics for binary vector (representing data from a single category)
}
\details{
This function differs from \code{catf} in that \code{catf} expects a categorical variable, while this expects a binary one (or one that can be treated as binary). Note that this function will \emph{not} work with factors or string vectors, while \code{catf} \emph{can} work with binary variables (simply producing an undesired entry for \emph{both} 0 and 1).

Also note that the default behavior is to \emph{not} include NA's in the denominator when calculating percentages. I think this is the right answer for most applications (and this matches the behavior of \code{catf}), but the option is available to change as needed.
}
\examples{
catf(ChickWeight$Diet)
catf(ChickWeight$Diet, which_in_paren = "n")

}
\seealso{
\code{\link[=table]{table()}}, \code{\link[=catf]{catf()}}
}
\author{
Michael Floren
}
