% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_factor.R
\name{ord_factor}
\alias{ord_factor}
\title{Order Factor}
\usage{
ord_factor(x, by = "count", decreasing = NULL, na.rm = TRUE, ...)
}
\arguments{
\item{x}{The factor vector to be reordered}

\item{by}{The method by which to reorder. Currently "count" and "alphabetical" are supported}

\item{decreasing}{The order to put factors in. NULL by default, with separate defaults for count (TRUE) and alphabetical (FALSE).}

\item{na.rm}{Should \code{NA}'s be removed?}

\item{...}{Additional arguments passed to factor.}
}
\description{
Orders factor levels based on count or alphabetical order
}
\details{
Orders factor levels based on count or alphabetical order
}
\examples{
#Note that while entries in each of the variables are the same, the order of the levels has changed
ord_factor(ToothGrowth$supp, decreasing = FALSE, by="alphabetical")
ord_factor(ToothGrowth$supp, decreasing = TRUE, by="alphabetical")

#this is most apparent when looking at tables of these (the order of presentation is different)
table(ord_factor(ToothGrowth$supp, decreasing = FALSE, by="alphabetical"))
table(ord_factor(ToothGrowth$supp, decreasing = TRUE, by="alphabetical"))

}
\seealso{
\code{\link[=factor]{factor()}}
}
\author{
Michael Floren
}
