% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel_end.R
\name{relevel_end}
\alias{relevel_end}
\title{Relevel Factor to End}
\usage{
relevel_end(x, ref, ...)
}
\arguments{
\item{x}{The factor vector to be reordered.}

\item{ref}{The reference level, to be placed at the end}

\item{...}{Additional arguments (currently not used).}
}
\description{
This function relevels a factor level to the end of the factor, rather than the beginning as in the typical \code{relevel} command.
}
\details{
This is a copy of the current relevel command, with an edit at the end to change the referenced factor level to come at the end rather than the beginning.
}
\examples{
levels(factor(ChickWeight$Diet))
levels(relevel_end(factor(ChickWeight$Diet), "1"))

}
\seealso{
\code{\link[=relevel]{relevel()}}
}
\author{
Michael Floren
}
