% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_stars.R
\name{sig_stars}
\alias{sig_stars}
\title{Significance Stars}
\usage{
sig_stars(p, breaks = NULL, labels = NULL, ...)
}
\arguments{
\item{p}{The p-value to calculate stars for}

\item{breaks}{The intervals that you're interested in doing stars for (defaults to 0, 0.001, 0.01, 0.05, 0.10, 1)}

\item{labels}{The labels for each of the interval spaces}

\item{...}{Other info passed to cut}
}
\description{
Based on p-values, output a specific number of stars
}
\details{
Uses cut to correctly specify p-values. Can handle multiple entries.
}
\examples{
sig_stars(.023)
sig_starts(.0023)

}
\author{
Michael Floren
}
