% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_contf.R
\name{sub_contf}
\alias{sub_contf}
\title{Continuous Descriptives By Subgroup}
\usage{
sub_contf(x, subgroup, names = TRUE, n = TRUE, ...)
}
\arguments{
\item{x}{The vector with data on the continuous variable to be used in calculating the descriptives.}

\item{subgroup}{The vector listing out the categorical variable to use in making subgroups}

\item{names}{Would you like the results to be named?}

\item{n}{An argument to pass to \code{contf}: would you like the sample size to be included? Included as an argument so that there isn't confusion with the names argument (no doubt due to my sloppy programming)}

\item{...}{Additional arguments to pass to \code{contf}}
}
\description{
This function calculates continuous descriptives of a variable broken out by subgroups
}
\details{
This is essentially an \code{aggregate} wrapper for \code{contf}.
}
\examples{
sub_contf(ChickWeight$weight, ChickWeight$Diet)
sub_contf(ChickWeight$weight, ChickWeight$Diet, names=FALSE)

}
\seealso{
\code{\link[=aggregate]{aggregate()}}, \code{\link[=contf]{contf()}}
}
\author{
Michael Floren
}
