% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_text.R
\name{wrap_text}
\alias{wrap_text}
\title{Wrap Text}
\usage{
wrap_text(x, width)
}
\arguments{
\item{x}{The character string/vector of text to be wrapped}

\item{width}{The max width that should be allowed (number of characters)}
}
\description{
This is a general text wrapping function similar to \code{label_wrap} from \code{scales}.
}
\details{
This follows closely with the \code{label_wrap} function from the \code{scales} package, but due to the funciton design is able to be used in a wider variety of contexts. Note that this simply inserts line-breaks at appropriate places to break up the string.
}
\examples{
ex_text <- "this is a very long text string"
wrap_text(ex_text, width=50)
wrap_text(ex_text, width=10)
cat(wrap_text(ex_text, width=10))

}
\seealso{
\code{\link[scales:label_wrap]{scales::label_wrap()}}
}
\author{
Michael Floren
}
